+++ src/AIEngineeringWorkspace.App/Infrastructure/AppInfo.cs
@@ -5,8 +5,16 @@ namespace AIEngineeringWorkspace.Infrastructure;
 internal static class AppInfo
 {
     public static string DisplayVersion
-        => Assembly.GetExecutingAssembly()
-            .GetCustomAttribute<AssemblyInformationalVersionAttribute>()?
-            .InformationalVersion
-            ?? "development";
+    {
+        get
+        {
+            var version = Assembly.GetExecutingAssembly()
+                .GetCustomAttribute<AssemblyInformationalVersionAttribute>()?
+                .InformationalVersion
+                ?? "development";
+
+            var metadataSeparator = version.IndexOf('+');
+            return metadataSeparator >= 0 ? version[..metadataSeparator] : version;
+        }
+    }
 }